+++ common/ASC.Core.Common/Billing/TariffService.cs
@@ -969,7 +969,7 @@ public class TariffService(
         return customerInfo;
     }
 
-    public async Task<string> TopUpDepositAsync(int tenantId, long amount, string currency, bool waitForChanges = false)
+    public async Task<string> TopUpDepositAsync(int tenantId, decimal amount, string currency, bool waitForChanges = false)
     {
         var portalId = await coreSettings.GetKeyAsync(tenantId);
